# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/skills/{agent}/{skill}": {
      "get": {
        "operationId": "SkillsController_getSkill",
        "parameters": [
          {
            "name": "agent",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "skills"
        ]
      },
      "put": {
        "operationId": "SkillsController_updateSkill",
        "parameters": [
          {
            "name": "agent",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "skills"
        ]
      },
      "delete": {
        "operationId": "SkillsController_deleteSkill",
        "parameters": [
          {
            "name": "agent",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "skills"
        ]
      }
    },
...
  },
  "info": {
    "title": "XIAOCHUANG Backend API",
    "description": "V2.0 unified backend for web and admin clients",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}